+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.0] - Unreleased
+### Added
+- New command *Reset layer names*.
+- New button to select the text layer while editing the variables.
+
+### Changed
+- The plugin has been refactored using [Create Figma Plugin](https://yuanqing.github.io/create-figma-plugin/).
+- The UI style is similar to Figma.
+
 ## [0.1.4] - 2021-09-15
 ### Added
 - 2 additional Google Spreadsheet urls to merge copies from different sources.
@@ -28,10 +37,12 @@ project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Replaced proxy url.
 
-## 0.1.0 - 2021-06-29
+## [0.1.0] - 2021-06-29
 First version
 
+[0.2.0]: https://github.com/marketgoo/figma-copies/compare/v0.1.4...HEAD
 [0.1.4]: https://github.com/marketgoo/figma-copies/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/marketgoo/figma-copies/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/marketgoo/figma-copies/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/marketgoo/figma-copies/compare/v0.1.0...v0.1.1
+[0.1.0]: https://github.com/marketgoo/figma-copies/releases/tag/v0.1.0